Cast column embeddings to FP32 in TabICLExplainer
When the base TabICL estimator runs inference under AMP, the trunk
returns FP16 column embeddings. The value head loaded from a FP32
training checkpoint then raises "expected scalar type Half but found
Float" on the first query. Cast at the explainer boundary so AMP
settings on the trunk and training precision of the heads can differ.

+++ src/tabicl/sklearn/explainer.py
@@ -323,7 +323,7 @@ class TabICLExplainer(BaseEstimator):
             f"Unexpected column-embedding shape {tuple(column_embeddings.shape)}; "
             f"expected (1, {h_filtered}, {model.embed_dim})"
         )
-        self._column_embeddings = column_embeddings.detach()
+        self._column_embeddings = column_embeddings.detach().float()
 
     def _compute_endpoints(self) -> None:
         """Cache predictive_sufficiency_ and predictive_necessity_."""
